alsa_sound: Fix for short break in music when touch tone plays
-When touch tone plays for the first time while music playback is
going on, a short break in music happens.
-In checkDeviceMuteStrategies whenever there is a change in device
on an output, tempMute is set. However the same tempMute is used to
decide whether to mute an already existing session on a different output
irrespective of the current active device on that output. Due to this
mute and unmute happens even when there is no real change in device on
a particular output which might cause a short break in playback.
-Add check to ensure that active streams are temporarily muted only
in case of a change in device on that particular output.

CRs-Fixed: 495009
Change-Id: Icbc2afa4516f700b170c1f6801135cc5ca61fbfe